# Suppport more scenes for fused_fast_ln