<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Others
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Suppport more scenes for fused\_fast\_ln
fused_fast_ln_kernel now can support the scenes bellowing：
- the number of column is 768 or 4096；
- the bias_ptr is not nullptr

Performance promotion：
- Fused OP: 

|        config                              | old   | new   | speedup |
| ------------------------------------ | ----- | ----- | ------- |
| bsz=1, in_seq_len=128, max_dec_len=8 | 10.01 | 7.45  | 1.34x   |
| bsz=4, in_seq_len=128, max_dec_len=8 | 11.30 | 8.21  | 1.38x   |
| bsz=1, in_seq_len=60, max_dec_len=20 | 22.27 | 17.07 | 1.30x   |
| bsz=4, in_seq_len=60, max_dec_len=20 | 23.05 | 17.54 | 1.31x   |

- Model：The average performance of Transformer model is improved by 1%~1.5%