@ZzSean 2022-04-26
> 1. 在PR描述里补充当列数为768和4096时该op的性能提升情况。
> 2. 该fuse kernel的寄存器使用量很多，会不会有溢出风险？

寄存器使用情况相较于修改前最多会多出Vec bias[LDGS]，LDGS最大也是4，且block中的线程数也是128，因此应该不会有溢出风险